Minor changes after comments on PR
