+++ doc/src/index.rst
@@ -17,6 +17,7 @@ MPI is used.
    hubs.rst
    spokes.rst
    extensions.rst
+   mmw.rst
    pysp.rst
    ef.rst
    secretmenu.rst
