+++ mpisppy/confidence_intervals/mmw_ci.py
